# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/register": {
      "post": {
        "operationId": "AuthController_userRegister",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRegisterDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully Registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/otp": {
      "post": {
        "operationId": "AuthController_sendOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendOtpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully Registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "AuthController_getCurrentUser",
        "parameters": [],
        "responses": {
          "200": {
            "description": "current user info",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "auth"
        ]
      }
    },
    "/users/password": {
      "post": {
        "operationId": "UserController_updatePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "update current user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "users"
        ]
      }
    },
    "/users/update": {
      "patch": {
        "operationId": "UserController_updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "update current user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "users"
        ]
      }
    },
    "/upload": {
      "post": {
        "operationId": "MediaController_uploadFile",
        "parameters": [
          {
            "schema": {}
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "multipart/form-data": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/CreateMediaDto"
                  },
                  {
                    "properties": {
                      "file": {
                        "type": "string",
                        "format": "binary"
                      }
                    },
                    "type": "object"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaDto"
                }
              }
            }
          }
        },
        "tags": [
          "upload"
        ]
      }
    },
...
  },
  "info": {
    "title": "API",
    "description": "### REST\n\nRoutes follow Richardson level 3.\n\n<details><summary>Detailed spec</summary>\n<p>\n\n…(your route descriptions)…\n\n</p>\n</details>",
    "version": "v1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "UserRegisterDto": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string",
            "minLength": 6
          },
          "phone": {
            "type": "string"
          },
          "otp": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        }
      },
      "RoleType": {
        "type": "string",
        "enum": [
          "USER",
          "ADMIN"
        ]
      },
      "UserDto": {
        "type": "object",
        "properties": {
          "uuid": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "username": {
            "type": "string",
            "nullable": true
          },
          "role": {
            "allOf": [
              {
                "$ref": "#/components/schemas/RoleType"
              }
            ]
          },
          "email": {
            "type": "string",
            "toLowerCase": true,
            "nullable": true
          },
          "avatar": {
            "type": "string",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "SendOtpDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "description": "phone number for sending otp"
          }
        },
        "required": [
          "phone"
        ]
      },
      "UpdatePasswordDto": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string",
            "minLength": 6,
            "description": "New password",
            "example": "Str0ng!Pass123"
          },
          "old_password": {
            "type": "string",
            "minLength": 6,
            "description": "Old password (optional)",
            "example": "OldPassword123"
          }
        },
        "required": [
          "password"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "first_name": {
            "type": "string"
          },
          "last_name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "toLowerCase": true
          },
          "avatar": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "MediaType": {
        "type": "number",
        "enum": [
          1,
          2
        ],
        "description": "1 : image -----------  2: video"
      },
      "CreateMediaDto": {
        "type": "object",
        "properties": {
          "type": {
            "description": "1 : image -----------  2: video",
            "allOf": [
              {
                "$ref": "#/components/schemas/MediaType"
              }
            ]
          }
        },
        "required": [
          "type"
        ]
      },
      "MediaDto": {
        "type": "object",
        "properties": {
          "uuid": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "fileExtension": {
            "type": "string"
          },
          "mediaType": {
            "type": "number"
          },
          "fileSize": {
            "type": "number"
          },
          "url": {
            "type": "string"
          }
        },
        "required": [
          "fileExtension",
          "mediaType",
          "fileSize"
        ]
      },
...
    }
  }
}